Minor cleanup and move some code
