+++ src/Titan.ECS/Scheduler/IRunner.cs
@@ -1,6 +1,8 @@
+using Titan.ECS.Worlds;
+
 namespace Titan.ECS.Scheduler;
 
 public interface IRunner
 {
-    static abstract void Run(ref ECS.Scheduler.Scheduler scheduler, ref World.World world);
+    static abstract void Run(ref ECS.Scheduler.Scheduler scheduler, ref World world);
 }
