+++ src/Titan.ECS/Scheduler/Runner.cs
@@ -1,14 +1,15 @@
 using Titan.Core;
+using Titan.ECS.Worlds;
 
 namespace Titan.ECS.Scheduler;
 
 internal unsafe struct Runner : IApi
 {
-    private delegate*<ref Scheduler, ref World.World, void> _run;
+    private delegate*<ref Scheduler, ref World, void> _run;
     public static Runner Create<T>() where T : IRunner =>
         new()
         {
             _run = &T.Run
         };
-    public void Run(ref Scheduler scheduler, ref World.World world) => _run(ref scheduler, ref world);
+    public void Run(ref Scheduler scheduler, ref World world) => _run(ref scheduler, ref world);
 }
