+++ src/Titan.ECS/Scheduler/SystemDependency.cs
@@ -1,5 +1,5 @@
 using System;
-using Titan.ECS.World;
+using Titan.ECS.Worlds;
 
 namespace Titan.ECS.Scheduler;
