+++ src/Titan.ECS/Scheduler/SystemsInitializer.cs
@@ -4,18 +4,18 @@ using Titan.ECS.Components;
 using Titan.ECS.Entities;
 using Titan.ECS.Events;
 using Titan.ECS.Systems;
-using EntityFilter = Titan.ECS.Entities.EntityFilter;
+using Titan.ECS.Worlds;
 
 namespace Titan.ECS.Scheduler;
 
 public readonly unsafe ref struct SystemsInitializer
 {
     private readonly SystemDependencyState* _state;
-    private readonly World.World* _world;
+    private readonly World* _world;
 
-    internal SystemsInitializer(ref SystemDependencyState state, ref World.World world)
+    internal SystemsInitializer(ref SystemDependencyState state, ref World world)
     {
-        _world = (World.World*)Unsafe.AsPointer(ref world);
+        _world = (World*)Unsafe.AsPointer(ref world);
         _state = (SystemDependencyState*)Unsafe.AsPointer(ref state);
     }
 
@@ -34,7 +34,6 @@ public readonly unsafe ref struct SystemsInitializer
 
     public ReadOnlyResource<T> GetReadOnlyResource<T>() where T : unmanaged, IResource
     {
-        //NOTE(Jens): should there be a destiction between local and global resources? 
         _state->ReadOnlyResources.Add<T>();
         return new(_world->GetResourcePointer<T>());
     }
@@ -48,7 +47,6 @@ public readonly unsafe ref struct SystemsInitializer
     public ReadOnlyStorage<T> GetReadOnlyStorage<T>() where T : unmanaged, IComponent
     {
         _state->ReadOnlyComponents |= ComponentId<T>.Id;
-
         return new(_world->GetComponents<T>());
     }
 
@@ -62,7 +60,6 @@ public readonly unsafe ref struct SystemsInitializer
         => new(_world->GetResourcePointer<T>());
 
     public EntityFilter CreateFilter(in EntityFilterConfig config) => _world->CreateEntityFilter(config);
-    public void GetEntities(in EntityFilterConfig filter) => _world->CreateEntityFilter(filter);
 
     public void RunAfter<T>() where T : unmanaged, IStructSystem<T>
         => _state->RunAfter.Add<T>();
