+++ src/Titan.ECS/Systems/SystemDescriptor.cs
@@ -1,5 +1,5 @@
 using Titan.ECS.Scheduler;
-using Titan.ECS.World;
+using Titan.ECS.Worlds;
 
 namespace Titan.ECS.Systems;
