+++ src/Titan.ECS/Worlds/IdGenerator.cs
@@ -1,6 +1,6 @@
 using System.Threading;
 
-namespace Titan.ECS.World;
+namespace Titan.ECS.Worlds;
 
 internal readonly struct IdGenerator<T>
 {
