+++ src/Titan.UI/UIButton.cs
@@ -1,5 +1,5 @@
 using Titan.ECS.Entities;
-using Titan.ECS.World;
+using Titan.ECS.Worlds;
 
 namespace Titan.UI
 {
