+++ src/Titan.UI/UIComponent.cs
@@ -1,6 +1,6 @@
 using System.Numerics;
 using Titan.ECS.Entities;
-using Titan.ECS.World;
+using Titan.ECS.Worlds;
 using Titan.Graphics;
 using Titan.UI.Common;
