+++ src/Titan.UI/UIManager.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using Titan.ECS.Entities;
-using Titan.ECS.World;
+using Titan.ECS.Worlds;
 using Titan.Graphics.Rendering.Text;
 
 namespace Titan.UI
